fix: fix font switching failed for the first time on a brandnew machine
FontSources are invalid on a brandnew machine, which causes swithing
font for the first time will end up with same lastPath and currentPath.
So this will be considered font is not changed.

Deleting assigning the value to lastpath in CheckFontSource and only
assigning the value in IsFontChanged will fix this issue.

Signed-off-by: wangxhhh <9450ave@163.com>